pyproject.toml
[project]
name = "autoclassifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "google-genai>=1.2.0",
    "polars>=1.22.0",
    "rich>=13.9.4",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "wandb>=0.19.7",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
